Refactor consolidation complete. Needs testing and adjustments.
