--- r_scripts/r_scripts.Rproj
@@ -1,14 +1,0 @@
-Version: 1.0
-ProjectId: e09313a0-53d7-4604-a829-331c299b7b87
-
-RestoreWorkspace: Default
-SaveWorkspace: Default
-AlwaysSaveHistory: Default
-
-EnableCodeIndexing: Yes
-UseSpacesForTab: Yes
-NumSpacesForTab: 2
-Encoding: UTF-8
-
-RnwWeave: Sweave
-LaTeX: pdfLaTeX
