--- r_scripts/test_account_data.csv
@@ -1,4 +1,0 @@
-match_id,account_id,team,hero_id
-34947428,22854130,0,1
-34947428,163176965,0,52
-34947428,394672146,0,3
