--- r_scripts/test_hero_match_data.csv
@@ -1,27 +1,0 @@
-hero_id,wins,losses,matches,total_kills,total_deaths,total_assists,total_net_worth,total_last_hits,total_denies,total_player_damage,total_player_damage_taken,total_boss_damage,total_creep_damage,total_neutral_damage,total_max_health,total_shots_hit,total_shots_missed,hero_pickrate,win_percentage
-1,25,47,72,473,573,656,2363548,10207,414,1701609,2058429,318975,1674588,2020657,194641,161759,105204,38.3,34.72
-2,58,48,106,594,806,1223,3693331,17404,599,2798361,2668097,761123,3221685,3558250,283132,151176,107829,56.38,54.72
-3,38,49,87,640,553,1051,2918379,10231,681,2284475,1459884,440183,2105332,1076982,186367,94506,75129,46.28,43.68
-4,52,30,82,570,557,927,2852753,11411,420,2447627,2549517,587917,2329326,1953378,270466,50463,24144,43.62,63.41
-6,32,39,71,540,571,833,2337473,9126,469,1736170,2382540,421779,1816859,1436347,204416,199858,129965,37.77,45.07
-7,43,49,92,655,552,852,3210371,14448,682,2373691,1996836,1023666,2477297,2856832,238796,344500,250920,48.94,46.74
-8,34,20,54,273,379,723,1873417,7643,412,1270406,1553881,514482,1419833,1333552,154067,154193,113905,28.72,62.96
-10,64,54,118,707,809,1753,3929556,14428,833,3066686,2728211,692099,2800694,1889474,321576,172432,148314,62.77,54.24
-11,22,26,48,170,359,641,1384829,4623,259,716477,1377526,163109,949485,406470,128732,29122,22476,25.53,45.83
-12,28,25,53,197,339,804,1633135,5190,427,886765,1727079,234903,1077420,574212,165974,31895,22208,28.19,52.83
-13,62,60,122,974,829,1229,4162388,17681,970,3005531,2722096,884342,2795846,3589627,300167,398074,251046,64.89,50.82
-14,35,39,74,507,441,1032,2349886,8709,365,1596766,1555314,396002,1944099,1047481,190328,33751,18697,39.36,47.3
-15,48,62,110,756,847,1335,3478646,13602,673,2707370,2911973,491745,2631852,1838487,311160,219555,180223,58.51,43.64
-16,33,31,64,481,437,674,2110729,8490,382,1545135,1786326,414602,1589702,1131670,197070,334239,293748,34.04,51.56
-17,31,60,91,623,588,1053,3015535,12017,479,2798181,1566417,391444,2830598,1538736,197732,86170,43858,48.4,34.07
-18,41,26,67,372,449,1021,2160680,8385,338,1334710,2258334,457969,1539618,1110103,248173,205498,185906,35.64,61.19
-19,51,63,114,927,841,987,3754388,15109,752,2827194,3485965,564900,2990707,2345582,344217,314456,209962,60.64,44.74
-20,36,35,71,324,409,858,2289115,10027,559,1559638,1783181,717668,1956275,1335559,187490,214257,172931,37.77,50.7
-25,48,52,100,700,773,1116,3357819,13884,475,2761848,2851944,890358,2899782,2100232,307467,110921,72813,53.19,48.0
-27,56,71,127,993,993,1307,4306455,18035,545,3450597,3860921,672257,3456384,3297690,317747,229446,114198,67.55,44.09
-31,89,76,165,1328,1058,2274,5505638,19923,885,4553117,3844958,757800,3985403,2822561,424512,160836,131499,87.77,53.94
-35,30,38,68,328,377,845,2171064,8650,351,1483123,1685713,338443,1877889,937691,186366,49999,38652,36.17,44.12
-50,70,41,111,878,710,1444,4012029,16626,745,3642284,2114776,727088,3372146,2604031,246763,292544,209363,59.04,63.06
-52,48,46,94,663,568,1119,3151590,12507,697,2582324,2143320,754121,2550259,1748189,257780,103004,56515,50.0,51.06
-58,38,21,59,543,423,528,2035330,8506,451,1537641,1482330,755714,1336880,1758827,151279,244091,177590,31.38,64.41
-60,16,20,36,209,228,442,1104360,3765,165,745212,857370,179556,784485,352578,92103,21927,20773,19.15,44.44
